Added augmented system info to Python
I added all of the new augmented system diagonstic variables to the
Python state.  I also fixed a bug where we didn't have the g_x variable
reported correctly in the state.  As a side note, it's nice that Python
has this built-in documentation, but this is somewhat redundant to our
comments and manual.  We need a better way to keep all three consistent.