Added documentation for the new diagnostics
I added the documentation for the new diagnostics to the manual.  I also
ended up adding documentation for a a few extra missing diagnostics as
well for things like the vector space diagnostics.  Anyway, it should
all be there and mostly correct.  There still a bit too much duplication
of documentation in different places that need to be consolodated
probably using some kind of metaprogramming technique.