Cleaned up truncated Krylov solvers
I cleaned up the truncated Krylov solvers to hopefully make more sense.
As part of this, there are two big changes.

First, I removed the trust-region scaling operator.  Basically, I never
really liked it and I'd not been using it for some time.  A long, long
time ago I tried using it to better help cut off the trust-region solve
when the inequality constraints became too close to the boundary.  Now,
since our inequality constraints are of the form h(x)>=0 where h : X ->
Z, there wasn't a very straight forward way to inject this information.
Typically, people end up using some sort of trick with inv(diag(X)) when
there are bound constraints on all variables.  Because the codomain of h
is Z, we can't exactly do this, so I used h'(x)*Linv(h(x))h'(x), which
did get us into the correct space.  In practice, it didn't work well and
I seem to recall a problem where the algorithm never converged.  The
reason for this is that the operator above isn't always full rank.  For
example, image the situation where we had a single inequality constraint
with many variables.  Then, the operator above is actually a rank-1
operator, which means it had a nullspace.  Since it had a nullspace, it
was possible, and did happen, that we never hit the trust-region bound,
which is bad.  Now, that doesn't mean that some sort of scaling wouldn't
be helpful, but I still think that doing this through the inner product
would be a better solution.  When it comes time for that, I'll revisit
the code.

Second, I actually had a small mistake with the NewtonCG code and the
truncated Newton method.  Basically, when we detect negative curvature,
we really should back off to the last step in order to maintain a
descent direction.  If this happens on the first iteration, then just
use the steepest descent direction.  Before this fix, I took the
direction of negative curvature and added a unit amount of it to the
current direction.  I never really noticed any problem with it, but it
was probably not right.

Anyway, that caused a bunch of rebasing on the examples.  Nothing major,
but some minor.  There's a couple of differences with the SOCP examples
between the MATLAB/Octave and C++ versions, which isn't great, but
hopefully these later algorithmic fixes will resolve that.

Finally, I futz around with a bunch of variable names that may keep me
more sane.  This is sort of a never ending cycle, but at least what I
have now matches my current set of notes.