Fixed a potential problem with instability
I fixed a potential problem with instability on the first iteration.
There, we want to move in the steepest descent direction.  Before, I had
the code just take this direction as is.  However, if we have a
trust-region method, we do need to truncate it first.  That's what this
code fix did.