Added a rough draft of the new interior point
I added in a rough draft of the new interior point scheme.  Basically,
the new idea is to better safeguard the interior point during the Krylov
method in order to insure that the Krylov method produces a strictly
feasible solution with respect to the inequality constraints.  Then, we
modify the inequality multiplier after we do our globalization.  In the
past, there was some difficulty where we weren't reducing the model
problem when taking into account the modification of the inequality
multiplier, something that CG should have been guaranteed to do.  In
addition, the safeguard has also been added to the quasinormal step,
which had no knowledge of the inequality in the past.  Finally, I
modified how and when we update the interior point parameter mu in order
to better address some numerical difficulties that arise when mu becomes
small.

Now, there's a whole bunch of code in this check-in that needs to be
eliminated.  I added second trust-region for the inequality constraint
step, which I didn't really like.  I also added a bunch of schemes for
selecting and updating the inequality multiplier, which I ultimately
didn't think worked well.  I'm including them in this check-in for
documentation sake in case I or someone else decides to mess with them
later.  There's an interesting zoo of how, why, and when things like the
multipliers can be calculated.  Someday, I'll document all of them, but
this checkin should give a good idea of many of the permutations.

I also found some other auxiliary bugs that I probably should have
checked in separately, but didn't.  There was a bug in the
golden-section search where the function evaluation wasn't correctly
returned.  That's been fixed, which messes with many of our tests.