Cleaned up interior point code, eliminated options
I cleaned up the interior point code and eliminated a number of options
that are oscentisbly deprecated.  This includes different centrality
strategies and interior point methods.  At this point, why hopefully
have the right one and I never really used the other strategies or
interior point methods that often.  Mostly, it was code that I had to
maintain, but didn't like.  In addition, I eliminated some of the other
pieces like the secondary trust-region radius.

At this point, the examples are kind of messed up still, so we need to
clean things up further.