Changed when inequality multiplier step calculated
I changed when the inequality step was calculated to before we do our
globalization.  This is what is was two check-ins ago and it appears to
work much better.  Now, what's odd about this location is that the
inequality multiplier step dz depends on dx.  When we compute dz prior
to globalization, we don't scale dx by alpha0 from the globalization.
Rather, we use dx raw.  For whatever reason, this appears to work much
better.  For a trust-reigon method, it doesn't matter if we compute dz
here or slightly later like we did in our last check-in.