Fixed examples and added alpha_x_qn
I fixed the simple_infeasible_inequality example.  Basically, the
quasinormal step was running into the boundary, which isn't good.  The
fix was just to increase the starting mu, so that we didn't bang into
the boundary right away.  In order to better detect this, I added a new
variable alpha_x_qn, which indicates how much we cut the quasinormal
step with respect to the safeguard.  If this number is small, then we
know we have problems.

Now, the output is a little messed up at the moment.  Basically, I want
this in the diagnostics, but I had to add the variable to the
unconstrained code so that everyone could see it.  I display this
variable in the inequailty constrained diagnostics, but really I only
want to do so when we have a fully constrained problem.  I probably
ought to move things around, so this is the case.

There's still something up with the MATLAB SDP code.  The example runs,
but the number of iterations is pretty far off from the C++.  That's a
little fishy, so investigation is necessary.