Added scheme to increase mu if necessary
I added in a heuristic that lets mu increase if we take small steps.
Right now, I just hard code this as 0.1, but it probably ought to be a
parameter.  Anyway, if we detect that a small step is taken, then the
code increases mu and resets z to what we choose for z initially.
Namely, z = mu L(h(x))^(-1) e.  The idea behind this is that we try to
force ourselves back onto the central path, but one that's easier to
obtain.

Note, originally, I tried this without reseting z and one of my problems
ended up increasing mu in an unbounded way.  That's bad and there may
still be some risk of this happening, so be mindful moving forward.