Fixed safeguard bug in CD and tangential step
I fixed two separate bugs.  The first was a bug in truncated-CD code
where we didn't properly safeguard things when we detected negative
curvature.  Basically, I didn't scale the direction properly, so it
turns out that we returned values that weren't feasible.  Second, I
added a safeguard search to the tangential step.  Note, part of the
reason for this big mess of a code change was to make the tangential
subproblem return feasible steps.  Now, in perfect precision, the
tangential step generated from this would also be feasible.  With
inexactness, this may not be the case.  As such, we need to do a
separate search and safeguarding of the tangential step in order to
guarantee that our combined step dx=dx_n+dx_t is feasible.