Moved everything to absolute tolerances
This is somewhat tangential to the branch, but I moved everything to
absolute tolerances and added a new variable eps_kind, which allows us
to move back and forth between absolute and relative tolerances for
optimization stopping conditions.  For Krylov stopping conditions,
things remain relative.  Anyway, this has been a major source of
irritation for me me and others since doing things like starting with a
feasible solution on an equality constrained problem means that we can
never satisfy that criterion.  I'm doing this now because some of my
test problems I end up setting things like eps_grad and eps_constr to
big values like 1e5 because they start so low and then we never reduce
mu.

In theory, we should be able to pick and choose what we want to be
relative and what we want to be absolute.  For the moment, it's all or
nothing, but that may change later.