Added diagnostics for normal and tangential steps
I added the norms of the normal and tangential steps to the output with
msg_level >= 3.  Basically, there's an issue when the norms are
radically different because the algorithm converges slowly.  For
example, when the normal steps are much smaller than the tangential, the
normal step will get us to optimality, but the tangential step pulls us
off.  This isn't exactly the same as the Maratos effect since we're not
rejecting the steps.  Plus, the augmented Lagrangian merit function
should be immune to Maratos.  However, it still means we converge
slowly.  Anyway, a rescaling of the objective separate from the
constraint appears to fix this problem.  Maybe there's a way to automate
the rescaling or use some other trick, but that'll be a task for later.