Fixed penalty parameter computation
I fixed a problem with the penalty parameter computation.

In the past, I had some problems with the algorith because the safeguard
for the interior point method wouldn't guarantee that the tangential
problem gave a reduction.  Basically, I cut the step and then dealt
with it in globalization.  That proved problematic, which necessitated
this branch of the code.  Now, barring inexactness, we can guarantee
reduction in the tangential and quasinormal steps.  However, before this
change, I had this safeguard that would not update the penalty parameter
when the predicted reduction was negative.  I seem to recall that this
may have helped, or not, but I thought I did at the time.  The problem
with that is that we need to adjust the penalty parameter in order to
guarantee a positive predicted reduction.  Basically, since we know we
reduce the pieces of the model that correspond to the quasinormal step
and the tangential step, we know that we can increase the penalty
parameter enough to guarantee a positive predicted reduction.  In other
words, we know that

    < W (grad L(x,y) + H dx_n), dx_t_uncorrected >
        + .5 < H dx_t_uncorrected,dx_t_uncorrected > < 0

from the tangential subproblem and

    || g(x) ||^2 - || g'(x)dx_n+g(x) ||^2 > 0

from the quasninormal step (unless we're feasible and then dx_n=0).
What we don't control directly in the predicted reduction is:

    < grad L(x,y), dx_n > + .5 < H dx_n,dx_n >
        + < dy , g'(x)dx_n+g(x) >

which may make things negative.  Now, there's probably some theory that
says that these things are bounded, etc, but the interior point cuts our
steps with the safeguard, so that makes things funny.  The bottom line
is that we can just elevate rho enough to force the predicted reduction
to be positive and that's what we're more properly doing now.