Removed truncated MINRES
I decided to remove truncated MINRES from our list of options.
Basically, I think that MINRES has some really nice features and can be
made to give very stable solves.  At the same time, it problem it solves
isn't exactly what we need in optimization algorithms.  Specifically, CG
solves the problem

0.5 <Ax,x> + <b,x>

whereas MINRES solves the problem

|| Ax + b ||^2

where, in the context of optimization, A is our Hessian and b is our
gradient.  Now, CG has the benefit that our truncated scheme is
guaranteed to give us a descent direction whereas MINRES is not.
Furthermore, CG guarantees descent for our trust-region model whereas
MINRES does not.  In the past, I compensated for some of these issues
with MINRES by flipping the trial step when we didn't get a descent
direction.  And, for the most part that worked.  However, with the
safeguards, we can't do this anymore.  Specifically, we're doing a check
in order to insure that dx is safe and that doesn't mean that -dx is
safe.  In theory, we could do more safeguarding on the negative
directions, but that would likely really complicate the code and it's
not clear to me if that would give a better algorithm.  Anyway, this
becomes a major issue with equality and inequality constraints since
when we obtain feasibility with respect to equality constraints, we
depend on a reduction in the tangential subproblem in order to guarantee
a positive predicted reduction.  Since we can't guarantee that with the
safeguards, at least easily, we end up with negative predicted
reduction, which leads to bad penalty parameter calculations, which
leads to the algorithm failing.  Certainly, I could keep around the
MINRES code for unconstrained problems.  To be sure, I have seen MINRES
give significantly faster solutions.  However, due to how I integrated
all of the codes, constraints or not, that would more complicate things
than I'd like and create more code to maintain than I'd like.  Finally,
I never integrated the loss of orthogonality (and lack of a perfect
projector code) into MINRES like I did CG.  It's possible, but more code
to create and maintain.  Frankly, this change will make my life easier.