Changed terminology back to CG from CD
I changed our terminology back from conjugate direction to conjugate
gradient.  It's a small change, but it's been irritating for me to
remember and type.  Basically, the origin of the name system came in
from the inexact composite step SQP algorithm.  There, the
preconditioner may change between iterations due to inexactness, so we
take truncated-CG and over orthogonalize against all previous
directions.  Since we're not just using the last Krylov vector, we can
call it conjugate directions.  At the same time, some people view
conjugate directions as an algorithm where we already have a bunch of
directions computed and make our search direction conjugate to them.
Basically, we could technically decouple conjugate direction from a
Krylov method and still have a conjugate direction method.  Anyway,
that's not what we do at all.  This is a Krylov method that really is
CG, but one where we forcefully orthogonalize things due to problems
with inexactness and roundoff error.  Even on unconstrained problems, it
makes sense to overorthogonalize if we can spare the memory because,
frankly, things become unstable pretty quickly.  Anyway, the name change
is a convenience to me and hopefully others who kept wondering what CD
versus CG meant.