Readded log-barrier method
I reintroduced the log-barrier method for first-order algorithms.  More
specifically, we now default to the log-barrier method whenever we don't
solve a linear system involving the Hessian, so technically we could
have a first order method, but as long as we solve a linear system,
we'll do primal-dual.  Really, this is just for inequality constrained
problems where we do something like steepest-descent or BFGS.  For any
trust-region problems or full constrained, we'll do primal dual.

In short, the primal-dual algorithm works much better.  Certainly,
there's probably a better way to do a log-barrier method, and perhaps
I'll figure that out and integrate it later.  However, in no uncertain
terms, if we can use the primal-dual method, we should use the
primal-dual method.  Really, the issue is that the primal-dual code
doesn't integrate properly into methods that don't do these linear
system solves since an integral part of the primal-dual code is this
Hessian modification.  The Hessian and gradient modification combined
make sure that we eventually migrate to the central path, which we use
for modifying our stopping conditions.  What was happening with these
first-order methods is that this wasn't happening and then values
migrated off into infinity or something bad.  Anyway, in these cases, if
we just force a log-barrier method, which we can essentially do by
ensuring that z = mu h'(x)*L(h(x))^(-1)e, the algorithms stabilize and
actually start to solve the inequality constrained problem.  That being
said, I find the convergence to be slow.  As such, there's probably a
trick that I'm missing, but at least the algorithm is technically going
in the right direction and the safeguards should mean that we stay
interior to the bound.  Honestly, if we only need a few digits of
convergence, it's not too bad.  Mostly, I think it takes too long to hit
the square root of machine precision accuracy, which may be due to
ill-conditioning, or possibly a bad algorithm.  Either way, it's
something that we can improve upon and we're better off in this
circumstance that just telling everyone to not do it.  That being said,
use the primal-dual algorithm with a Newton-type method.  It's better.