Modified ctest test names
I modified some of the ctest test names to disclose a more complete path
for the test.  This was already the case for the JSON tests, but now
it's more uniform for everything else.  Mostly, this help me debug
things a little quicker because I couldn't keep track of what test was
where without turning on verbose mode with ctest.