Fixed error function in MATLAB/Octave
I fixed the error function in the MATLAB/Octave code.  Basically, MATLAB
doesn't really like it that we call the error function when catching an
error.  In truth, we don't really need to since we throw a MATLAB error
automatically when something bad happens, so we really only need a print
statement.