Added correct computation of norm_dxtyp
I fixed the comptuation of norm_dxtyp to be the size of the first step.
Originally, this was altered in case we had error on our first
iteration.  As such, we start with this value being norm_dxtyp as a
safeguard.  However, after we calculate dx, we really should have
updated it and now we do.