Added qn_stop information and herustics
I added a new variable called qn_stop, which stands for quasinormal
stopping condition.  Basically, it tells us whether or not we actually
took the Cauchy, dogleg, or Newton points and whether or not we hit the
safeguard.  Outside of providing information about the structure of the
problem, we use this information when we decide whether or not to
increase the size of the trust-region.  Basically, if the actual vs
predicted reduction is good and the quasinormal problem hits the
trust-region (specifically zeta delta of it), then we now increase the
trust-region radius.