Changed TR to a truncated dogleg method
I changed the unconstrained and inequality constrained problems to a
truncated dogleg method.  In terms of performance, it's mostly a wash.
There's a couple of problems where we get better performance, but I'm
not completely convinced that the new scheme will always win out.  That
being said, conceptually I like it better because I really hate
recomputing the same thing over and over again, which is what happens
when we restart truncated-CG and simply cut the trust-region.
Basically, the first couple of iterations are probably the same unless
we cut down to the steepest descent point.