Modified truncated-CG variable names
I modified the truncated-CG variables names from 'krylov' to 'trunc'.
Originally, I used the name krylov this or that because we had both CG
and MINRES as truncated solvers for the optimality system.  Since I
removed MINRES due to problems with the safeguards, we only use CG now.
Now, I could rename everything to 'cg', but we have a bunch of nonlinear
CG algorithms in the code, so that would confuse things.  I could leave
the name as krylov, but that's also confusing because the augmented
system solves use GMRES and that's also a Krylov method, one in which we
probably want different settings.  In addition, the SQL vector space
uses the implicitely restarted Arnoldi method to find eigenvalues for a
line search, which is also a Krylov method and I have different
parameters for that.  Really, we need something that's unique to the
solver that we use for doing our optimality solves.  I did think about
using 'opt' for optimality solver, but then there's the variable
opt_stop that carries the overall optimization algorithm stopping
condition.  I couldn't figure out a better name for that one, so we're
stuck with 'trunc' for now.  I also changed trunc_rel_err to trunc_err.
It's still a relative error, but I got tired of dealing with 'rel' and
all of the other parameters went to the err suffix with the
absolute/relative error change.

Also, I removed all of the messaging code for truncated CG as well.
It's been turned off for some time as I wasn't fully happy with how the
output was interleaved.  I've a couple of ideas for getting more
information from truncated CG and displaying it with diagnostics, but
that'll come later.