Modified globalization tracking
At the core of these changes was modifying how we track the
globalization progress.  Now, we keep track of exactly what
globalization iteration we're on, be it trust region or line search.  We
also keep track of the total number of globalization iterations.  In
addition, we have a cap on the number of such iterations.  If the
algorithm exceeds that number, then we exit from the algorithm with a
new error GlobalizationFailure.  This is hopefully another safeguard to
help us exit in case something goes terribly wrong.

Because we're tracking globalization iterations, I removed
rejected_trustregion as it's been deprecated.

Because of the new dogleg method setup, we'll eventually retreat to the
Cauchy point, so I removed history_reset, which was a safeguard to
insure this happened eventually.  This may have helped some other
problems, but I never noticed it.

Because we adding a new stopping condition, I changed the name of the
enumerated type from StoppingCondition to OptimizationStop to have the
name better align with TruncatedStop and QuasinormalStop.

Because I got tired of typing it, I changed the linesearch prefix to ls.
This also aligns more with the tolerance eps_ls.

There was also a bug where we set alpha, alpha0, and delta to align
despite what algorithm we use.  This allows us to change algorithms more
quickly.  Anyway, at one point in the past, the line search would go out
to 2 alpha0.  This is no longer the case; it only goes out to alpha0,
but this syncronization was messed up.

I also fixed this ongoing issue with reporting the truncated CG
iterations.  Basically, with our current setup, we run truncated CG at
most once per optimization iteration.  We want to report its
performance, such as iterations and error, but we didn't have a good
mechanism for doing this once.  Now that we track what globalization
iteration we're on, we only output this information when glob_iter is 1.
Basically, once per iteration.  If not, we output our blank separator,
which is currently '.'.  This is a little nonideal since it would be
nice to have this information everytime we complete a new optimization
iteration, but that's also awkward and redundant.  Anyway, having the
new glob_iter variable along with the dogleg stuff is what allows this.