Updated CMake to auto set C++ version
I updated the CMake setup to auto set the C++ version to C++14 when
required and C++98 when not.  Basically, we only need C++98 for jsoncpp
when we build it.  Anyway, it looks like the parameter that controls
this is CXX_STANDARD, which I think was introduced on CMake version 3.1.
The documentation for CMake is currently at 3.1.3 for the 3.1 release,
so we'll use that as our new minimum.  The documentation has been
updated accordingly.