Added documentation for direct cmake configuration
I added documentation for direct cmake configuration that doesn't use
ccmake.  Most of the time, ccmake should be easier because it has a
guided menu system.  However, I keep getting requests to build things in
a script, so I might as well document it.  This actually exposed a
problem with building LAPACK after such configuration, so it was a good
exercise.