Added jsonlab dependency and SQL serialization
I added a jsonlab dependency and then used it for the SQL serialization.
In short, I could code the serialization for the SQL vector space by
hand, but jsonlab can do this better than I can at the moment, so it's
easier just to use that.  In truth, we should probably just use jsonlab
for all of the MATLAB serialization as it works pretty well.  Likely,
this will be part of the next design phase, so I wanted to add the
dependency now to get used to it.

By the way, I have sort of an odd hack to install jsonlab from the
archive for the user.  I'm sure there's a better way to do this, but it
works for now.