Rough draft of some safeguard improvements
I added code to help improve the safeguarding on both truncated-CG and
the quasinormal step.

Previously on truncated-CG, if we exit due to negative curvature or
hitting the trust-region, we wouldn't take the step unless the last
iterate was feasible.  Now, even if the last step was infeasible, as
long as the next step is feasible, we'll take it.

On the dogleg method, the situation was similar.  In addition,
previously, we exited if the Cauchy point violated the safeguard.
Really, the Newton point may not, so we continue computing until we
violate the trust region or complete.  Then, we have to assess the
safeguard in a very similar manner to CG.  Meaning, even if the Cauchy
point isn't feasible with respect to the safeguard, we need to find the
Newton point because that point may be and we need to check.  I also
added a check that we handled in the quasinormal step manipulator to
exit if the Cauchy point brings us to optimality.  If this happens, we
can't hit the stopping tolerance on the augmented system solve for the
quasinormal step.

There were a couple of test problems that sped up with the change, so I
updated them.

By the way, the quasinormal stopping condition is sort of messed up
right now and needs to be revised.