Fixed bug in actual reduction for equality constr.
I fixed a pretty bad bug in how the equality constrained problems
computed the actual reduction.  Basically, the merit function at the
current step was calculated with the equality multiplier at the next
step.  Basically, this occured because the function modifications
implicitely look at the multipliers to do their computations.  I updated
the multiplier too early when calculating the merit function at the
current step.  This is a weakness in how the modifications work, so I
need to come up with some better schemes to prevent these bugs.