Added tracking for failed augmented system sovles
I added a new set of variables for failed augmented system solves both
the state and to the output.  In short, we look at augsys_failed_total
to tell us when we're failing augmented system solves.  Here, failure
means that we can't meet the target tolerance.  In any case, this is not
the end of the world, but is generally bad and normally means that
there's a problem with the preconditioner.  Really, convergence means
that we need to meet these tolerances, so things may fail when this
doesn't happen.