Changed names of GMRES preconditioners
I changed the names of the GMRES preconditioners from Ml_inv and Mr_inv
to B_left and B_right, respectively.  Basically, they're just operators
and we don't exactly need an inverse to be useful.  A lot of this
terminology is a hold over for how the algorithms are most often
presented.