Added checks for nonsymmetry and nonprojector
I added some new checks for nonsymmetry and when we have a nonprojector.
Before, this was all rolled into the orthogonality check, which only
occured for equality constrained problems.  Per our comments for the new
code, we now have O where

O_ij = <B ri, rj> / (|| B ri || || rj ||)    i!=j
     = <B ri, ri> / || Bri ||^2 - 1          i=j, proj_check=true
     = 0.0                                   i=j, proj_check=false

From this matrix, we check the following in order

1. || Diag(O) || > epsilon

   B is not a projector. Needed in the composite step SQP method.
   Note, when B is a projector we have || Bri ||^2 = <Bri,Bri> =
   <B*Bri,ri> = <B^2ri,ri> = <Bri,ri>, so we really should get 0.

2. || O - O' || > epsilon

   B is not symmetric.  Hard requirement, but users give strange
   things.  Basically, <Bri,rj> = <ri,B*rj> = <ri,Brj> when B is
   symmetric.  Hence, O should really be symmetric.

3. || O || > epsilon

   Vectors are not orthogonal.  In theory, the
   preconditioned/projected residuals are orthogonal to each other
   and the diagonal elements should be 0 due to the discussion
   above.  As such, O should be zero, but it becomes not the case
   when one of the checks above fails or, barring that, we don't
   have orthogonality in the preconditioned residuals.  We're using
   a variant of Gram-Schmidt, so this will happen eventually due to
   numerical error.

As far as epsilon, we pick the magic number 0.5.  There's probably a
better number as this is sort of one of those, "Well, that looks
good," numbers.

I've added tests for nonprojector and nonsymmetry as well.

I also added a missing file, spaces.h, that the previous check-in or so
required in order to compile

Finally, I added a preconditioner to the basic truncated-CG unit test.
I could add tests forever, but in the interest in brevity, I added an
appropriate preconditioner to the basic setup.