Added ability to over orthogonalize
I added the ability to overorthgonalize during truncated-CG.  At the
moment, we use modified Gram-Schmidt for orthogonalization.  This is
more stable that classical Gramm-Schmidt, but it'll eventually break
down.  In order to make it more robust, we can play a couple of games
such as those with Householder reflectors.  Maybe someday it'll come to
that, but for the moment, that seems to require some odder algebras to
implement.  A simple alternative is just to run the orthogonalization
again, which is an old idea and one that works as well as Householder
reflectors most of the time.  When I originally learned this trick, I
heard it called the DGKS correction step.  I've also heard it called
iterative modified Gramm-Schmidt (IMGS.)  At the end of the day, it's
really just the orthogonalize again trick.

I changed the name of trunc_orthog_max to trunc_orthog_storage_max to
better fit with trunc_orthog_iter_max variable added to implement the
over orthogonalization described above.  This'll break some old code and
I'm sorry for that.