Added test for objective increase
I added a test to insure that we exit properly when the CG objective
increases when truncated-CG.  In theory, this shouldn't happen, but it
sometimes does if we can't meet our stopping tolerances.  This is bad in
the sense that we require the CG objective to decrease monotonically in
order to guarantee model decrease for trust-region methods.  As such, it
makes sense to do a hard test and exit whenever this occurs.