Added unit test for negative curvature in tcg
I added a test that helps validate the negative curvature pieces of the
truncated-CG code.  In addition, I combined the loss of orthogonality
and regaining orthogonality unit tests.  Basically, they've mostly the
same problem setup, so they should be in the same file.  The same issue
came up in the negative curvature unit tests, trust-region vs none, and
I squished it all together, so it made sense to do the same for the loss
and regain of orthogonality examples.