Added test for Cauchy point and negative curvature
I added a test for when we detect negative curvature on the first
iteration.  Really, the trust-region method should work pretty much the
same way, but a line-search method should return the steepest descent
direction unscaled.  The idea is that even when we have an uncooperative
Hessian, we should always have descent.