Reworked truncated-CG code and added diagnostics
I reworked the truncated-CG code to, hopefully, make it easier to
understand what's going on.  Essentially, I've continued to add features
and diagnostics to the code, which tend to hide the actual algorithm.
To combat this, I started adding some different bookkeeping code to make
the diagnostics and storage of terms more concise.  This came to head
now because I'm in the middle of adding some diagnostic checks to the
linear operator A and I felt like there was too much boiler plate code
in the routines and that the checks were a distraction.  Now, that being
said, the current set of stopping conditions needs to be updated because
we now have things like symmetry checks both on A and B.  I also needs
some tests for these failures.  The code should also be cleaned up and
some of the lambda functions moved to be real functions.  Finally, I'll
need to go through this process with the safeguards, but that'll come
later.