Added test for detecting nonsymmetric operators
I added a test for detecting nonsymmetric operators.  This required some
relabeling of the old truncated-CG stopping conditions to make room for
the new case.  Now, like nonsymmetric preconditioners, we need to
actually being storing vectors for this test to work;
trunc_orthog_storage_max needs to be big enough.  Also, there's a
reasonable chance that the negative curvature or objective increase
stopping conditions may trigger first.  I don't know of a good way to
avoid this because all of these tests work on tolerances and it may take
time for one to go up vs another.  In addition, if we have negative
curvature, we can't really store the old preconditioned directions
either because we generally store the normalized value, which is the
square root of the quanitity we check for negative curvature.  As in,
the normalization would be a nan.  Really, as long as we quit
appropriately, we can mostly continue with the algorithm.  That said,
it's nice to know when something is obviously wrong like a nonsymmetric
operator as quickly as possible.