Reworked truncated-CG and moved out some functions
I reworked the truncated-CG code and moved out of the function some of
the functions that I recently added to help with the diagonstics.
Certainly, some that require closures dependent on locally available
information remain.  Anyway, this sort of makes things uglier in the
sense that we lose type deduction outside of lambda functions in the
current version of C++.  My guess is that they'll eventually extend it
to everything, which will allow these functions to be simplified at a
later time.