Reworked safeguard code in truncated-CG
I reworked the safeguard code again in truncated-CG and added a test
that insures that the offset is feasible.  In theory, the feasibility of
the offset with respect to the safeguard should always happen, but odd
things do occur.  Anyway, most of this is in preparation of adding a
bunch of tests that validate each of the cases where the safeguard code
becomes active.