Added a unit test for three different VS
I added a unit test for three different vector spaces.  In truth, this
unit test isn't really designed to be run, though it can be done safely,
but rather to check that we have a type safe program with respect to our
vector spaces.  Our code is pretty heavily templated, but we really want
to separate the vector spaces X, Y, and Z from each other.  As long as
the types are compatible, C++ is perfectly happy to use X operations on
Y vectors, but that's a bug.  This unit test sets up three incompatible
vector spaces in order to better find these places.  As an auxiliary
benefit, it insures that we're also not calling the copy constructor,
which we souldn't be doing currently.