Fixed bug in QN stopping condition, started unit
I fixed a bug in the quasinormal stopping conditions.  Basically, I'd
left in the old code for computing the stopping condition, which was
wrong.  That affected one of the unit tests, which has been modified to
account for the changes.

Outside of that, I started putting in some unit tests to better validate
the quasinormal step.  This first one validates that we exit properly
when the cauchy point brings the linearized feasibility to 0.  This is
important because if we don't exit, the augmented system solve should
fail because we can't hit the stopping tolerances.