Fixed bug when quasinormal step at local min
I fixed a bug where the quasinormal step is essentially at a local min,
so we don't have a good idea of where to step.  Basically, we compute
the Cauchy point by solving the least-square problem

min_dx 0.5 || g'(x)dx + g(x) ||^2

which has a gradient of

g'(x)^*g'(x)dx + g'(x)*g(x)

If we start with an initial guess of dx=0, then the steepest descent
direction is -g'(x)*g(x).  Then, we do a line-search with this direction
in the least-squares objective above to find the Cauchy point.  Now,
when g'(x)*g(x)=0, the gradient above is zero and we're stuck at a local
min.  Therefore, we now set qn_stop to be LocalMin, and take a zero
step.  This occurs on constraints like

g(x,y) = 0.5 x^2 + 0.5 y^2 - 1

where we start at (x,y)=0.  Now, there's a pretty good chance that the
tangential step is going to move us off this point, so it may not be a
huge deal.  However, if we already forced the Lagrangian optimality
condition to be zero, then we're really stuck.  Hopefully, this doesn't
happen too often, but if it does, we can now diagonse why.  I also added
a test that makes sure that we throw the right exit criteria.