Added checks for when the quasinormal Newton fails
I added checks and tests for when the quasinormal Newton step fails.  In
theory, this could be for a variety of reasons and we probably ought to
separate those out later, but the easiest case to generate is when the
derivative of the equality constraints is rank-deficient.  In this case,
the augmented system solve will fail.

I added a check to GMRES to try and detect when something bad is
happening.  Right now, that check is when a new orthogonalized Krylov
vector is essentially zero.  When this happens, we exit GMRES and
correctly annotate that an augmented system solve failed.  At this
point, we now move back to the Cauchy point and exit the routine.
Finally, I added a unit test to make sure we verify this behavior.